requirements.txt
# For web.
django==4.2.3
djangorestframework==3.14.0


# [code_style]-[BEGIN]
# Pre-commit hook.
pre-commit==3.3.3
# Upgrade syntax.
pyupgrade==3.9.0
# Formatter. Apply codestyle.
black==23.7.0
# Check codestyle.
flake8==6.0.0
# Removes unused imports and unused variables.
autoflake==2.2.0
# [code_style]-[END]
